// src/lib/enums.ts
export enum CategoryCode {
  O = "O",
  P = "P",
  H = "H",
  C = "C",
  X = "X",
  M = "M",
  B = "B",
  K = "K",
}

export enum CategoryBase {
  Region = "Region",
  District = "District",
  Community = "Community",
  Village = "Village",
  Urban = "Urban",
  City = "City",
  DistrictCity = "DistrictCity",
  CitySpecial = "CitySpecial",
}

export enum KattogStructureType {
  Region = "region",
  District = "district",
  Community = "community",
  Settlement = "settlement",
}